Updates to About.php and Faq.php
Added actual readable text to Faq.php I'm still working on the css for
this. It works perfect in my test environment but not so much on the
live server.

I changed About.php back to the original design. I remember you
commenting on how you liked that design more. I tend to agree!